% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spatial_data.R
\name{load_spatial_data}
\alias{load_spatial_data}
\title{load_spatial_data Get the spatial data}
\usage{
load_spatial_data(
  name = "VG250_KRS",
  column = "SN_L",
  filter_by = NULL,
  crs = 25832
)
}
\value{
A data frame
}
\description{
load_spatial_data Get the spatial data
}
\examples{
data <- load_spatial_data(name = "VG250_KRS", column = "SN_L", filter_by = NULL, crs = 25832)

head(data)
}
